AdclusterServer/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=0.19.0,<1.0.0
sqlalchemy-utils>=0.38.0,<1.0.0
passlib>=1.7.4,<2.0.0
bcrypt==3.2.0
python-jose>=3.3.0,<4.0.0
python-multipart>=0.0.5,<1.0.0
websockets>=15.0.0,<16.0.0
aiofiles>=0.8.0,<1.0.0